UIBar Prototype
Started the UIBar. Still have to wire up the values to represent what they actually are in code. Still working on trying to make the elements of the bar fit the screen even when the screen is very thin. HorizontalLayoutGroups and the AutoSize component of TextMesh Pro Texts are not being kind to me, though.