StreamingAssets, working build, and UnityLogsViewer
Moved all of the game data to the StreamingAssets folder from the Data folder. This folder persists after a build which means XML files, SQLite databases, and AssetBundles can be loaded this way.

Got the build working on Windows. Still need to test MacOS and Linux.

Added the plugin UnityLogsViewer to the project which allows you to view the Unity logs from inside a build by making a circular motion with the mouse while holding it down in game. This is extremely useful to find any issues with the build easily.

I benchmarked the game's performance on my desktop and laptop.

On my desktop, the lowest my framerate got was ~200fps in Level1-1-AP. In some other areas, it was up at ~700fps.

On my laptop, the lowest my framerate got was ~70fps in Level1-1-AP. In some other areas, it was up at ~120fps.

My laptop is about as low-end as you can get, so these results are great. With a bit of optimization with the AI and making sure that no level is >75% the size of Level1-1-AP, we will be all set.

Keep the levels compact and fun, and scatter non-hostile areas throughout.