Added item saving, fixed camera in dialogue, and more
Implemented item saving so that now whenever the player adds/removes items from their inventory the most updated version of the inventory persists after the game is shut down and restarted.

Items that are placed in the world (i.e. dropped or opened from a chest) are saved to the database in the level that they are in. Whenever that level is asynchronously loaded, the items are loaded as well. If the item is picked up by the player, then it is removed from the database.

When items are instantiated into the world, they are given a unique id which describes them in the database. This is necessary so that the right items are added/removed from each level at runtime.

This took a bit of tinkering in the level loader where the newly loaded scene is set as active for a split second to load the items into the right scene. Without doing that, all loaded items would be a part of the persistent scene and if the level is unloaded, the items wouldn't be and then they could be reloaded over and over again.

Fixed the camera in dialogue to focus on the halfway point between the player and NPC/Sign. If the player is within 1.5 units of the NPC/Sign, they move away from them and then face them in order to create some personal space. If the NPC is flagged as able to turn, then it turns towards the player for the conversation.

Fixed crates so that they no longer jitter when moved. I set the rigidbody to interpolate which smoothed everything out.

Added a physics material to the player which removes all friction from them. They slide around a bit more while running (barely noticeable), but it makes it so that they cannot stick to walls which is nice. This also makes it so that crawling is never slower if you start motionless as opposed from rolling. I'll most likely add a bit of logic to slow the player down quicker so that they do not slide.

The camera zoom out is a little fast at the moment. I'm trying to find the balance so that the zoom speed is proportional to the distance zoomed.

The camera no longer follows the player once they start dying. If they fall in a pit, the camera stops following them once their death is initally triggered.

Added back a small dampening value to the cinemachine camera which smooths everything out. However, the reason it was disabled in the first place was for teleportation. When the player teleports, the damping would cause the camera to try and follow the teleport which was not ideal.

After entering a door or dying, I added a small delay which allows: the damping factor to be set to zero, the player to teleport, and then the damping factor to be set back to its original value which is a great workaround.

Added a Cinemachine group target component which allows for the camera to target multiple targets and fit them all in the screen. This can be used to focus on something important that is happening in the scene while still focusing on the player as well. A trigger will have to be set up to only focus when within a range, though.

Fixed the jumping that would happen if a menu is exited with spacebar. Now, I wait until the end of the frame to reenable input in those cases which makes it so the spacebar press does not register as a jump for that frame.

Fixed the text in the backpack and interact menus to all use the same text.

Fixed the interact bar to be a bit larger and now increase size with the text held within. The downside here is that when the screen is very small, it will overflow from the screen. However, the screen will most likely never be that small.

Changed the collapsing tile to use a poly collider instead of an edge collider. This makes it so the collapse can be triggered below the tile which prevents a rare bug of being trapped underneath one in crouch. Edited its physics shape to match its outline.

Added Zenject to the project. This is not being used yet, and it may never be, but I imagine putting it in here may motivate me one day to set it up so that we can properly unit test.

Got rid of CrossPlatformInput since it is not really needed. Our input system will do everything we need it to. Also, I got rid of the examples in the TextMesh Pro project to save a bit of space.

Added a FollowTarget component which is used by the Cinemachine group target alongside the player. Whenever we need to focus on something other than the player (like an NPC/Sign), the target's transform is set to the FollowTarget so that the camera tracks the middle of the player and target.

Adjusted the Dialogue and Backpack menus to not overlap the UIBar. Made them both a bit smaller vertically.

Verified that the Windows x86_64 build still works. I'm probably going to start building the game each time I make a commit.