Optimized the player, improved the UI, fixed UI element focus
The player's animations now prevent the player from interacting when they should not interact. Before, the interaction trigger was being enabled in some animations where it should not have been.

Adjusted the player's hitbox so that it no longer causes the player to jitter when walking over a group of unconnected tiles (like the collapsing tiles in Area1-1-AP). It took a bit of balancing the angle relative to the floor of the bottom of the poly collider on the player.

Adjusted the player's hitbox so that they can no longer stand on the one-pixel-wide side of a crate and jump up. Like with the previous paragraph, I slanted the bottom of the player's poly collider so that if they try to land on a thin gap they slide off immediately and never land.

Improved the dialogue box UI so that the choice container only appears if there is a choice to make. Otherwise, this box is set behind the actual dialogue container. In the event that there is no choice to make, a "continue" text appears at the bottom of the dialogue container where the player just presses the select button to move to the next dialogue segment. It has been animated as well.

Improved the interact container UI so that the collect all items button is as wide as all of the other choices in the container. The navigation is slightly improved as a result.

Fixed the UI element focus so that if the player clicks away from a button, it sets focus to any interactable selectable. This makes it so that if the player is not using a mouse to navigate the menus, there is always something to press.

Fixed the way that each menu disables focus of their UI element. Now, it sets the currently-selected UI element to null on exit so that the next menu can inherit the focus properly. Each menu with dynamic content (the inventory and interact containers) sets the focus element initially to the first dynamic element in the scrollrect.

Added the FMOD and University of Michigan–Dearborn logos so that they appear on the splash screen at the start of the game.

Verified that the Windows x86_64 build still works properly.