Added Rat and Dom, updated Backpack navigation, updated UI focus logic
Added Rat and Dom spritesheets. Set up the Rat as an enemy and Dom as an NPC that you can talk to in the castle.

The Rats center of mass will need to be centered in their spritesheets as right now when they turn around, they can clip through the map.

Updated the Backpack navigation so that it is much smoother now. Each set of buttons in the Backpack now have a set navigation so that you cannot access buttons that you should not be able to access.

Updated the UI focus logic so that if the player clicks away or focus is lost, focus is regained right away so that the keys or controller can be used to navigate.

Added a raycast blocker at the front of the canvas so that the mouse cannot be used in the menus. This is done because the navigation is now optimized for controllers/keyboard keys and the mouse does not behave well with things like scrollrects.