Added Slime, fixed UI focus issue, set NPC animations
Added the Slime and all of its animations. Set it as a prefab.

Fixed a UI focus issue which prevented the Event System from resetting focus for an element which is no longer active. A check is now done to see if the current element is inactive in the hierarchy.

Set the NPC animations for Dom, King, and General. They all now use the CharacterAnimator which they override. In dialogue, whenever the character is speaking, they enter the talk state and exit once the characters stop typing on the screen.