Sleeping, animations, input logic
Added sleeping to the player. If the user does not give any input for 30 seconds, the player will fall asleep and wake up upon input.

Organized a bunch of animations. Added the weapon animations to an overridable animator which each weapon overrides. Whenever a weapon is changed, the animator for that weapon can be equipped and then the material for the rarity can be assigned to that weapon.

Changed the input logic to allow for sleeping and make it work even when the player is in a menu.