Sleep sheet fix and component tagging
Added the updated player sleep spritesheet which fixes the issue of the dark outlines on it.

Added component tagging which replaces the tagging/layer system. Instead of using tags to define what an object is, empty MonoBehaviours have been added to the objects which describe them. For example, the ground tilemap has components standable which means that a character can stand on it, and blockable which means the player cannot uncrouch when blocked by it from above.

The current list of tags includes:

Standable - character can stand on it.
Blockable - player cannot uncrouch under it.
Player - player class.
Climbable - player can climb it.
Interactable - player can interact with it.
Attackable - character can attack it.

Some of these classes, like Attackable, Interactable, and Player, have existed previously.

The Trigger class has been revamped to no longer use tags and layers. Now, the Trigger abstract class has a template which the derived classes specify with one of the component tags. This template is used with a check to see if the object in the trigger has a component of the template type. If that check is not null, then it does.