Player info saving
Player information now gets saved and loaded. This includes information like the player's current health, maximum health, gold, spawn, and equipped weapon and armor.

The spawn gets set as the player passes through it, and each of the health, gold, weapon, and armor gets set as it changes.

The UIBar now updates the player's gold, location, and current quest.