Fixed two UI bugs
The InteractionMenu was allowing for enabled input while the menu was open in rare occasions (when spamming it open and closed). This has been fixed.

Fixed another issue where if the orthographic distance (change between the start and finish orthographic size) is very low, the camera shift lerp time will not allow for shifting since it is relative to it. The lerp time is now set to a minimum of a half of a second to allow for a shift even when the orthographic distance is low. This was an issue when you open a menu immediately after closing another.