Platforms
Platforms have been implemented.

From this commit, open the persistent level and add PlatformTest-SC. Ensure that the level loaders are disabled as well as orthozooms. Here is what each button does:

Orange: pressing this button moves the platform in a loop back and forth from its starting to ending location. These buttons can be set up to require an object on top of them or just be pressed once to activate.

Green: pressing this button moves the platform towards the starting location.

Yellow: pressing this button moves the platform towards the ending location.

The orange buttons do not play nicely with the yellow/green buttons because each are designed to be used separately. I'm just combining them for effect here.

I had to mess with the Character movement to get this to work properly. They now use Update for everything. This may have caused issues with the AI, but I doubt it.

Each platform must have its starting and ending locations set as well as their speed of travel. Keep the speed below 5 or it's a bit too fast for characters.

Platforms can be rotated 90 degrees and used as an obstacle with spikes.

Make sure that the bottom/side of the platform (whichever pushes up against the ground) has spikes on it so that it kills the player instead of clipping them through.