Check the log
Changes are as follows:

- Created the Ice tile, I made three of them all different kinds. One is
just a flat one with a bright layer on top so when we change the color
it looks like ice. The other two are the diagonal ones. ISSUE with them,
even though the slippery physics material is applied nothing happens to
the knighty when he is on it.

- Dissappering tiles are in the game. The timing on them needs to be
adjusted so it looks smoother, but they are properly working.

- Metal platoform was designed and properly added into unity

- I added the spikes in. Designed the sprite itself then created a
script that allows us to change how much damage is done if anything with
an attackable component lands on it. Note about the spikes colliders,
they are kind of inside the sprite itself just cause, I will change it
but I thought of it as I was writing this.

- I created a sleeping bag for Knighty and properly added it into the
game. I may go back and change how it looks because as of right now I am
not a big fan.

- Created a scene for the Ice level

- For the Wraith: I completed every animation except for Attack 2 , I
have just started that one. An issue I found with these like 2 minutes
ago is that the sprites themselves are a little above the ground, so I
am going to have to go back and lower them. I dont think we need a wind
up for the wraith, and if we do decide to switch it, All the animations
work perfectly off the idle animation.

- Removed Dapaidanon

- I feel like I forgot to explain something that changed.