Added the rest of the SQLite tables and queries
Added all of the necessary tables into the SQLite database. Created read/write queries to access each of the tables. The tables are updated via the observer pattern. Here are the set of tables:

Bosses
CharacterInfo
CharacterItems
Checkpoints
LooseItems
Openables
PlayerInfo
PlayerItems
Quests

Some of these still have to be wired in, but the SQLite side of things should be all set barring a design change.