NPC saving, event system, shop, dialogue actions
Added NPC saving and loading. Instead of existing within a scene, NPCs now are loaded with the scene (just like loose items). Each NPC has a prefab in its name in the Assets/Resources/Characters folder which is used to load them at their specified location in their specified level.

Added an event system which notifies the player of anything important that happens throughout the game such as an item pickup, location change, openable unlock, etc. It starts off on the right side of the screen, but if a menu shifts to that side, it will switch to the left side of the screen. Once an event has been in the first index for 2.5 seconds, it will start to disappear unless an identical event is triggered beforehand in which case its exit timer will be reset.

Added a shop as a dialogue action. When the player converses with a character who has this action, it will load up the shop which is very similar to the player inventory. The player can buy items from the character's inventory or sell items from their own inventory given that either party has enough gold for the transaction in either case. When done shopping, the dialogue is continued.

Added some dialogue actions like making the character disappear, shopping, giving gold, giving items, and taking gold. Not all of the dialogue actions have been implemented yet, however. You can try out a couple of them by talking to Dom in Castle1-AP.