Poseidon has Joined
Added the Water and the Boats

Water:
It is a prefab with a default background, we will have to scale
accordingly for the areas(shouldnt be an issue). The Desnity of the
water is 25 so when the player jumps on it it is still dense enough to
support that. Crates are way to heavy to be able to float (can probably
use that for a puzzle). The background on the water changes based on the
waters magnitude, it gives the illusion of movement.
WATER DOES NOT KILL AS OF NOW, I wanted to push this before I broke
anything.
Boat:
The boat is a basic collider with the boat image, Its mass is 9 so when
the player goes on it, it will become 10. This helps the boat stay above
water.  One issue is that the player slides off.

Also added a single spike so we can attach them to the metal platforms.