Almost done with the Main Forest Level
Had to stop because my mouse broke.
Also created an object looping script for the items in the water (boats,
crates, etc.) it seems to be working fine if the boats are moving right
to left but not the other way around, I will look into this.

Issues with the level currently:
-First Vertical platform doesnt move when the button is pressed.
- hatch in first area is not locked on start
- Water looks weird
- Dont really know how to make the platform puzzle in the third part.