Check the log
Added a speed variable to the OrthoShift script, this allows us to make
more dramatic zooms if needed.

Fixed the WaterBehavior

Changed how the first puzzle in Forest-MAIN works.

Also added a flipflop ability to the Water Magnitude: It will flip
ebtween negative and positive every time change. This allows us to make
more scripted water switches and will be helpful in the boss fight