Weapon/armor equipping, item data changes, healing, attack force change
Weapons and armor can now be equipped once picked up. Currently-equipped weapon and armor will be put back into the player's inventory. You can see which weapon/armor the player currently has equipped in the inventory as well.

Once equipped, the weapon's animator override controller will be applied so that the correct spritesheet will be used. Also, for both the armor and weapon, the correct material will be applied to match  (i.e. diamond, emerald, etc.).

Changed the item data to now include a name for the material used as well as the sprite referenced. This makes it so that multiple items can use the same material or sprites and allows for less items to be maintained in the resources folder.

Not all of the permutations of the weapons/materials have been added yet. The weapons/armor are still purely visual in terms of how they affect the game. I'll have to do some balancing to figure out how much damage the best weapon should do vs the worst weapon and how much damage reduction should be done for the armor.

Implemented healing where using a potion will restore health up to its strength.

Changed the attack force to be very tiny in all cases so that it feels less clunky. I think this is a happy medium between no force at all and too much force.