Validated build
Had to make a slight change to the WaterBehavior script in order to get the build to work. Any script that makes use of the UnityEditor namespace needs to be placed in an Editor folder. Since this script did not actually need the namespace, I just removed it.

The build now works. The player slides a bit more than I would like in the build and the zoom happens too quick for the regular menus.