Updated Goblin, more DialogueActions, added spear, fixed BoxCast, CharacterData, Materials, Attacking, UI improvements
Updated the goblin animations to include the new spritesheets.

Added the request and teleport dialogue actions. Characters can now be sent to different levels and locations within a level with a fade to black and back. Characters can also now request an amount of a certain item or of gold to continue a dialogue path. If the player does not have it, then a different path is traversed. A submenu within the dialogue menu was created to handle this.

Added the spear and all of its animations (just like how the sword and axe are set up).

Fixed the AI boxcast so that it is in the correct place now. Before, it was a box with its center at the center of the enemy and did not change directions with the enemy. Now, it changes directions and its side starts at the center of the enemy so that the enemy cannot see behind. Added a debug square to the world which visualizes the BoxCast.

The CharacterData has been changed to now include information related to attacks. Any attacks that a character may have are detailed in the CharacterDatabase XML file. This file is used for any Attackable now, including the crates and dummy. All attackables pull their health from this file at runtime. Still need to set up strength, agility, and weight to be used, however.

As mentioned above. CharacterAttack now makes use of the CharacterDatabase. The added benefit here is now the wind up times for enemy attacks and odds of an attack occurring are kept in the file instead of globally set. They can be set on an individual level now.

Changed the BackpackMenu and StoreMenu so that the loaded items now are eplicit in their navigation. There were certain scenarios where an item would navigate incorrectly upwards into a tab, but now since they explicitly navigate to each other it is no longer an issue. A coroutine had to be created to wait for previously loaded items to be destroyed before the new ones were loaded because they were being considered for the navigation linking uncorrectly.