Fixed AI, fixed damage knockback, enemy prefabs, and movement logic
Fixed the AI so that it now chases the player when it is supposed to. There was one issue that I noticed when an enemy finds the player in the bounds of a beacon, but I was not able to reproduce the issue.

Enemies now have a large trigger box in which, if the player stays inside after being detected, the enemies will follow the player. If the enemies are a certain distance away from the player, they will start to run towards them.

Fixed the damage knockback so that it works correctly now. I set the previous velocity of the attacked object in Attackables when the knockback force is applied to the updated velocity after the force. This makes it so that the updated force is the point from which the velocity decreases, not the other way around.

Updated all of the enemy prefabs with the sight box that detects the player. Also, I reduced the distance of the jump triggers because they were getting caught on stairs.

Changed the player movement logic slightly so that they do not fly out while attacking or dying. This effectively makes it so that the attack forces are entirely negligible.