Implemented dummy death animation, added input override function skeletons, added character profiles for ExplodingCrate and DestroyableTile
Implemented the Dummy death animation so that it can now be destroyed.

Added some input override functions in the InputManager so that the PlayerInput can make use of these functions. This allows for us to use our custom mappings at runtime which the Input class does not allow otherwise.

Added character profiles/data in the CharacterData XML file both with 1 vitality. They were using the Crate profile before, but it is better to separate them out.