Reverted Loading
The only thing that has to change in the SaveData file is the file variable which holds the name of the database that will be loaded for the game. That can be pulled from playerprefs after it is set which file is being loaded in the main menu. Everything from the main menu (options, etc.) will be set in playerprefs.