SaveData fix and partial Ogre implementation
I wanted to push a fix for the SaveData. Currently, it is setting a string "CurrentSave" on its construction which is temporary until the main menu sets this string on a game load.

Added the Ogre. In the process of setting up a Boss script and OgreBoss script which will be used to give behavior to all bosses and the ogre respectively. Boss is where a check will be done to instantiate the boss if it has not been defeated as well as setting the database entry to no longer spawn it. OgreBoss will contain a method for each animation which can be used to script it out. More in a bit.