Level loading fixes from main menu
Set up a couple of scripts to load the adjacent levels upon starting the game. This works from the main menu directly as well as after exiting to main menu from within the game.

Adjusted the player in the main menu to only have the idle animation as to prevent it from messing with the save data.