Project organization, golem import, and database update
Reorganized the project to clean it up a bit. Got rid of the AssetBundles which were not really doing anything important.

Imported the Golem and started laying out the scripts for the boss fights.

Updated the database to reference the string name of the bosses as opposed to an id.