Removed boss health variable and made a couple boss fixes
Forgot to remove this after I made the Boss script inherit from Attackable. The Attackable script has a currentHealth value which should be used instead. It is initialized with the values from the XML CharacterData sheet in the StreamingAssets folder.

Bosses can now take damage and upon being killed, the query sent to the database is correct where it was slightly wrong before.