Removed Zenject, moved physics materials, name conversion, princess boolean, quest XML setup
Removed Zenject since dependency injection isn't going to be used with this project. Moved the Unity Standard Asset physics materials to the Materials folder in a subfolder called Physics.

Added name conversion so that the dialogue/quests will say prince/princess, he/she, etc. depending on who the player is saving between the prince and princess. This is done by string replacement right before the text in question is loaded.

Added an isSavingPrincess boolean to the player info in the database and inside the player class depending on who the player is saving between the prince and princess. Initially, this value will be set from the main menu through player prefs, but after that it will be referenced from the database.

Added the quest data implementation. Each XML file under the Quest directory in the StreamingAssets folder holds a quest with a name, description, and set of steps. The first quest step should be 1. Quests will all start at step 0 where they are undiscovered. Once discovered, they will progress to step 1. After completion, they will be set to step -1 which is the completion state.

More to come later with wiring up data, journal, and map implementation.