Debug boolean fix
Added a debugging boolean to allow the game to be tested without having level loading enabled.

The GameData object in the Persistent-SC scene has a script called PersistentLevelLoad which initially loads the player and the surrounding levels. Since no levels can be loaded while disabled, the player gets stuck here. A boolean isDebug makes it so that this load does not happen if it is not desired.