NPC animations, quest+checkpoint saving, quest progression, become hostile, particle manager
Added all of the NPC animations and created a prefab resource for each to be loaded with the levels.

Added saving for quests and checkpoints. Quests are now stored at runtime (as well as persistently) with their current step. Checkpoints are stored on whether or not they have been reached yet.

Added quest progression through dialogue. When the ProgressQuest action is encountered, a selected quest will have its step changed to the specified value.

Created functionality which allows NPCs to become hostile. When the BecomeHostile action is encountered in dialogue, the hostile character will be removed from the database and start attacking the player. They are kept dormant until they start attacking.

Added a particle manager that spawns in particles in the form of sprite animations on the screen at the location of characters. For example, the healing particle spawns on the player once they use a health potion. The alert particle spawns on an enemy when they spot the player.

Will be working on the user interface for quests and fast traveling soon. The backend implementation is completed, but the front end needs to be exposed in the BackpackMenu still.