UI Update, button animation, pickup animation
All of the buttons on each user interface element now fade and scale when they are selected. This allows the user to have a better idea of where they are navigating on the screen at any given time.

Pickups now animate up and down when they are in the world.

The scrollboxes are now less jerky with their movement. Now, the vertical scroll position within the scroll box is lerped when set.

Map and journal UI will be up in a bit.