Quest and Map UI, Prince(ss), and various small fixes
Added the user interface for the journal and map. The journal allows you to see all of your active quests and select one to put on the UIBar. The map gives you a list of all of your unlocked checkpoints for the purpose of fast traveling.

The prince/princess has been added and will exist as a single gameobject. When loaded in, the prince/princess will have a script that figures out whether the prince or princess is being saved. Then, the animation controller of the correct character will be loaded.

Fixed the event system, interaction menu, journal, etc. to show the prince/princess's actual name instead of the @ symbol.

Added the death animation for one of the hostile NPCs.

Added the new quest and location logos to the UIBar.

Next, I will be wiring up the NPCs into the levels and adding the door unlock method  on a boss death. Then, the tutorial area.