Main menu and input updates
Updated the style of the main menu to make it fit more in line with the in-game UI. Consolidated everything into a single canvas and made it so that everything scales regardless of the resolution.

Reorganized the input mapping under controls. Now, the correct control pops up when selected. Some work still needs to be done to make it work for the axes for the gamepads.

I'm considering just leaving the options menu in the main menu since it is so quick to get to.