Some dialogue, level linking, bear fix, and more
Added some dialogue for the General, Sketch1, and signs.

Linked the abandoned village into Level1-1-AP. Linked the new tutorial area to the castle/start of game.

Fixed the bear so that it now spawns and works as it did before. Now, doors that unlock on a boss defeat use the BossDefeatedUnlockAction.

We now use a conversion to take something like {inputKey} in dialogue and convert it to the current key input.

Fixed an issue with the travel menu that was causing it to break when the player dies.

Fixed the hostile characters to be scaled 1,1,1 instead of 1/3,1/3,1/3 which made them 3 times larger on instantiation.

Fixed the way that quests populate the UIBar. If there is no active quest and a quest was updated, now it is set as the active quest.

Added Attacker, Robot, and Horseman to the CharacterDatabase. Fixed Bear in CharacterDatabase by giving it nonzero stats for strength, speed, and weight. Fixed all of the other characters by giving them a weight of 1.

Added a new training area which allows the player to learn the controls in a compact area.

Built the game for Windows x86 and x86_64, MacOS, and Linux x86 and x86_64. All builds succeeded but I was only able to validate that both Windows builds were working properly as of now.