Project reorganization, input bug fixes
Reorganized the project. Upgraded FMOD from 1.10.02 to 1.10.04 which fixes a couple of warnings in the log. Moved all of the FMOD files into the FMOD folder within the plugins folder. Rearranged the prefabs in the prefabs folder and scripts in the scripts folder so that they are less messy.

Fixed an input bug which was preventing the knight from falling asleep after 30 seconds. There was a missing if statement which was added in the PlayerInput script for crouch input checking.

Fixed another input bug which allowed the player to move as they are dying. This bug happens when the player loses input and later gains input after their death has been initialized. This creates a situation where the reenabling of the input happens when it shouldn't, so the logic for reenabling input has been modified to exclude situations where the player is dying.