Dialogue, quests, hooded guy, controller support
Added a lot of dialogue and some interleaving quests which will be expanded in a bit.

Added the hooded guy as the last NPC.

Added controller support. The exact button names are not displayed as of yet (just joystick button 0, 1, etc.) but that is an easy fix with the NameConversion class.

Important events now get written to a text file. This text file is to be parsed at the end of the game to review the decisions with the player.