Results scene, event mask fix
Added the results screen which loads the log file for the current save and moves through all of the choices made by the player in the game, analyzing each as it progresses. At the end, an overall score is applied to the player which gives them an overall idea of how they performed throughout the game and provides feedback accordingly.

Fixed the event mask to not display so that the far quarter of the screen is not lit up slightly.