Modified music, added sound effects, volume
Modified the music files in FMOD to loop.

Added some sound effects for navigation. These have their volume set upon scene instantiation and were definitely the hardest to implement since they had to be placed directly in the custom input module.

Figured out how to set the volume of events. It is one extra line of code, but as far as I know it has to be set on all events which is not that big of a deal.