Added the rest of the dialogue/quests, fixed some level issues, fixed climbing
Climbing was not working correctly because RigidBodyMovePosition does not scale in update for some reason. Had to add the AddForce to FixedUpdate and work with it there. Now, climbing is consistent across all framerates.

Added the rest of the dialogue and quests. Let me know if there are any issues in the way that they are carried out.

Fixed some level issues in Plains. Cave, Spooky, Swamp, Lava, and Ice still need to be fully playtested. There are some gamebreaking issues that prevent a full playthrough.