Credits, sounds, input icons, and platform fix
Added some items to the credits and expanded the width of them.

Lowered some sounds and moved the player death sound to be right when they die.

Removed the input icon pack as we are not using it.

Fixed platforms to check if charactermovement is null before applying velocity changes to them.