Dialogue fixes
Fixed some items in the dialogue which did not make much sense and also prevented forward progress with the game.

For the Squire and General, they are teleported to the Persistent level in their quest lines. However, they are removed from the level upon doing so which is a problem because they are never respawned in if the player completes the game in one sitting since the Persistent level is never reloaded. This has been fixed.