Removed saves, logs, set checkpoints false, and fixed new game issue
Removed all of the save databases and log files from the StreamingAssets folder.

Set all of the checkpoints to false so that you first have to unlock them to travel to them.

Fixed an issue upon making a new save file where it would overwrite a previous save. This now works correctly.