Final cleanup
Removed the build debug tool and removed save databases and log files.

In the previous commit, I changed some boss songs around, moved the exploding crate noise to a different spot in the script, removed the base flinch call from the golem script, reworked some of the control displaying in dialogue, and fixed another issue with how the general and squire moved levels.